requirements.txt
astroid==2.4.2
attrs==20.3.0
behave==1.2.6
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
dataclasses==0.8
glob2==0.7
google-api-core==1.25.0
google-auth==1.24.0
google-cloud-core==1.5.0
google-cloud-speech==2.0.1
google-cloud-storage==1.35.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
grpcio==1.35.0
idna==2.10
importlib-metadata==3.4.0
iniconfig==1.1.1
isort==5.7.0
lazy-object-proxy==1.4.3
libcst==0.3.16
Mako==1.1.4
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==20.8
parse==1.19.0
parse-type==0.5.2
pluggy==0.13.1
proto-plus==1.13.0
protobuf==3.14.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pylint==2.6.0
pyparsing==2.4.7
pytz==2020.5
PyYAML==5.4.1
requests==2.25.1
rsa==4.7
six==1.15.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.2
wrapt==1.12.1
zipp==3.4.0
